backend/requirements.txt
fastapi
uvicorn[standard]
passlib[bcrypt]
python-dotenv
PyJWT
email-validator
